Begin allowing score submission

+++ scripts/button.py
@@ -4,9 +4,9 @@ import aabb
 import common
 
 #Constant values for typical centered menu
-WIDTH  =  common.WINDOW_WIDTH  / 4
-HEIGHT =  common.WINDOW_HEIGHT / 10
-LEFT   =  common.WINDOW_WIDTH  / 2  - WIDTH / 2
+WIDTH  =  int(common.WINDOW_WIDTH  / 4)
+HEIGHT =  int(common.WINDOW_HEIGHT / 10)
+LEFT   =  int(common.WINDOW_WIDTH  / 2  - WIDTH / 2)
 
 def create(aabb, text, window, fill):
     x1 = aabb["x"]
